+++ src/plugins/Equalize.js
@@ -12,8 +12,7 @@
 var notSupportClamp = FILTER._notSupportClamp,
     CHANNEL = FILTER.CHANNEL, MODE = FILTER.MODE,
     FilterUtil = FILTER.Util.Filter,
-    A32F = FILTER.Array32F, stdMath = Math,
-    Min = stdMath.min, Max = stdMath.max;
+    stdMath = Math, Min = stdMath.min, Max = stdMath.max;
 
 // a simple histogram equalizer filter  http://en.wikipedia.org/wiki/Histogram_equalization
 FILTER.Create({
@@ -23,13 +22,13 @@ FILTER.Create({
 
     ,mode: MODE.INTENSITY
     ,channel: 0
-    ,factor: 1.0
+    ,factor: 0.0
 
     ,init: function(mode, channel, factor) {
         var self = this;
         self.mode = mode || MODE.INTENSITY;
         self.channel = channel || 0;
-        if (null != factor) self.factor = +factor;
+        if (null != factor) self.factor = (+factor) || 0;
     }
 
     ,serialize: function() {
@@ -52,17 +51,18 @@ FILTER.Create({
             r ,g, b,
             rangeR, rangeG, rangeB,
             cdfR, cdfG, cdfB,
-            f = self.factor,
-            t0, t1, t2,
+            f = self.factor || 0,
+            t0, t1, t2, v,
             i, l=im.length;
 
         cdfR = FilterUtil.histogram(im, CHANNEL.R, true);
         cdfG = FilterUtil.histogram(im, CHANNEL.G, true);
         cdfB = FilterUtil.histogram(im, CHANNEL.B, true);
         // equalize each channel separately
-        rangeR = f*(cdfR.max - cdfR.min)/cdfR.total;
-        rangeG = f*(cdfG.max - cdfG.min)/cdfG.total;
-        rangeB = f*(cdfB.max - cdfB.min)/cdfB.total;
+        f = 1 - Min(Max(f, 0), 1);
+        rangeR = (cdfR.max - cdfR.min)/cdfR.total;
+        rangeG = (cdfG.max - cdfG.min)/cdfG.total;
+        rangeB = (cdfB.max - cdfB.min)/cdfB.total;
         if (notSupportClamp)
         {
             for (i=0; i<l; i+=4)
@@ -70,9 +70,12 @@ FILTER.Create({
                 r = im[i  ];
                 g = im[i+1];
                 b = im[i+2];
-                t0 = cdfR.bin[r]*rangeR + cdfR.min;
-                t1 = cdfG.bin[g]*rangeG + cdfG.min;
-                t2 = cdfB.bin[b]*rangeB + cdfB.min;
+                v = cdf.binR[r]*rangeR;
+                t0 = ((f)*(v + cdfR.min) + (1-f)*(cdfR.max - v));
+                v = cdf.binG[g]*rangeG;
+                t1 = ((f)*(v + cdfG.min) + (1-f)*(cdfG.max - v));
+                v = cdf.binB[b]*rangeB;
+                t2 = ((f)*(v + cdfB.min) + (1-f)*(cdfB.max - v));
                 // clamp them manually
                 t0 = t0<0 ? 0 : (t0>255 ? 255 : t0);
                 t1 = t1<0 ? 0 : (t1>255 ? 255 : t1);
@@ -89,9 +92,12 @@ FILTER.Create({
                 r = im[i  ];
                 g = im[i+1];
                 b = im[i+2];
-                t0 = cdfR.bin[r]*rangeR + cdfR.min;
-                t1 = cdfG.bin[g]*rangeG + cdfG.min;
-                t2 = cdfB.bin[b]*rangeB + cdfB.min;
+                v = cdf.binR[r]*rangeR;
+                t0 = ((f)*(v + cdfR.min) + (1-f)*(cdfR.max - v));
+                v = cdf.binG[g]*rangeG;
+                t1 = ((f)*(v + cdfG.min) + (1-f)*(cdfG.max - v));
+                v = cdf.binB[b]*rangeB;
+                t2 = ((f)*(v + cdfB.min) + (1-f)*(cdfB.max - v));
                 im[i  ] = t0|0;
                 im[i+1] = t1|0;
                 im[i+2] = t2|0;
@@ -107,8 +113,8 @@ FILTER.Create({
         if (MODE.RGB === self.mode) return self._apply_rgb(im, w, h);
 
         var r, g, b, y, cb, cr,
-            range, cdf, i,
-            l = im.length, f = self.factor,
+            range, cdf, i, v,
+            l = im.length, f = self.factor || 0,
             channel = self.channel || 0,
             mode = self.mode;
 
@@ -137,16 +143,19 @@ FILTER.Create({
             cdf = FilterUtil.histogram(im, CHANNEL.G, true);
         }
         // equalize only the intesity channel
-        range = f*(cdf.max - cdf.min)/cdf.total;
+        f = 1 - Min(Max(f, 0), 1);
+        range = (cdf.max - cdf.min)/cdf.total;
         if (notSupportClamp)
         {
             if (MODE.GRAY === mode)
             {
                 for (i=0; i<l; i+=4)
                 {
-                    r = (cdf.bin[im[i+channel]]*range + cdf.min)|0;
+                    v = cdf.bin[im[i+channel]]*range;
+                    r = ((f)*(v + cdf.min) + (1-f)*(cdf.max - v));
                     // clamp them manually
                     r = r<0 ? 0 : (r>255 ? 255 : r);
+                    r = r|0;
                     im[i] = r; im[i+1] = r; im[i+2] = r;
                 }
             }
@@ -154,17 +163,19 @@ FILTER.Create({
             {
                 for (i=0; i<l; i+=4)
                 {
-                    r = (cdf.bin[im[i+channel]]*range + cdf.min)|0;
+                    v = cdf.bin[im[i+channel]]*range;
+                    r = ((f)*(v + cdf.min) + (1-f)*(cdf.max - v));
                     // clamp them manually
                     r = r<0 ? 0 : (r>255 ? 255 : r);
-                    im[i+channel] = r;
+                    im[i+channel] = r|0;
                 }
             }
             else
             {
                 for (i=0; i<l; i+=4)
                 {
-                    y = cdf.bin[im[i+1]]*range + cdf.min;
+                    v = cdf.bin[im[i+1]]*range;
+                    y = ((f)*(v + cdf.min) + (1-f)*(cdf.max - v));
                     cb = im[i+2];
                     cr = im[i  ];
                     r = (y                      + 1.402   * (cr-128));
@@ -186,7 +197,8 @@ FILTER.Create({
             {
                 for (i=0; i<l; i+=4)
                 {
-                    r = (cdf.bin[im[i+channel]]*range + cdf.min)|0;
+                    v = cdf.bin[im[i+channel]]*range;
+                    r = ((f)*(v + cdf.min) + (1-f)*(cdf.max - v))|0;
                     im[i] = r; im[i+1] = r; im[i+2] = r;
                 }
             }
@@ -194,7 +206,8 @@ FILTER.Create({
             {
                 for (i=0; i<l; i+=4)
                 {
-                    r = (cdf.bin[im[i+channel]]*range + cdf.min)|0;
+                    v = cdf.bin[im[i+channel]]*range;
+                    r = ((f)*(v + cdf.min) + (1-f)*(cdf.max - v))|0;
                     im[i+channel] = r;
                 }
             }
@@ -202,7 +215,8 @@ FILTER.Create({
             {
                 for (i=0; i<l; i+=4)
                 {
-                    y = cdf.bin[im[i+1]]*range + cdf.min;
+                    v = cdf.bin[im[i+1]]*range;
+                    y = ((f)*(v + cdf.min) + (1-f)*(cdf.max - v));
                     cb = im[i+2];
                     cr = im[i  ];
                     r = (y                      + 1.402   * (cr-128));